fix: Domain -> Dto 로직 responseDto로 이동
