+++ module-presentation/src/main/java/com/depromeet/member/facade/MemberFacade.java
@@ -4,7 +4,6 @@ import com.depromeet.member.domain.Member;
 import com.depromeet.member.domain.MemberGender;
 import com.depromeet.member.domain.vo.MemberSearchPage;
 import com.depromeet.member.dto.response.MemberSearchResponse;
-import com.depromeet.member.mapper.MemberMapper;
 import com.depromeet.member.port.in.command.SocialMemberCommand;
 import com.depromeet.member.port.in.usecase.MemberUpdateUseCase;
 import com.depromeet.member.port.in.usecase.MemberUseCase;
@@ -44,6 +43,6 @@ public class MemberFacade {
 
     public MemberSearchResponse searchByName(String nameQuery, Long cursorId) {
         MemberSearchPage memberSearchPage = memberUseCase.searchMemberByName(nameQuery, cursorId);
-        return MemberMapper.toMemberSearchResponse(memberSearchPage, profileImageDomain);
+        return MemberSearchResponse.toMemberSearchResponse(memberSearchPage, profileImageDomain);
     }
 }
